Remove fonts directory creation from ensureAssetsDirectories
This commit removes the creation of the `fonts` subdirectory within the `assets` directory.  The `ensureAssetsDirectories` function now only creates the `assets` and `images` directories.  The unit tests have been updated to reflect this change, verifying that the `fonts` directory is not created and throws an error if accessed.  This simplifies the asset directory structure and removes unnecessary directory creation.