Refactor: Pass `originalRect` as argument to `calculateInsertIndex`
The `calculateInsertIndex` function, used for drag-and-drop functionality in the spreadsheet, previously relied on accessing `columnDragState.originalRect`. This caused issues, as noted by the comment "// Always here!!".  This commit refactors the function to accept `originalRect` as an argument. This removes the dependency on the component's state within the function and makes the code cleaner and more maintainable.  The unnecessary dependencies in the `useCallback` hook are also removed.