Fix: Allow dropping assets at the end of the column
The `maxIndex` calculation in `ColumnDragCalculator` was incorrectly preventing dropping assets at the very end of the column.  The `Math.max` function was subtracting 1, resulting in an off-by-one error.  This commit corrects the calculation to allow dropping at the last position.  The comment has also been updated to clarify that the last position is now allowed.